Scroll improvement
This scrolls all the way to the top if the y-coordinate of the current field is equal or less than the y-coordinate of the first field. This is useful when theres a table at the top of the page. When scrolling up to the first row, the column headers will now be shown.