Background function improvements
Improved RTC synchronisation by periodically sending the time to the FC until we get the response confirming that the data was received and OK.

Improved rssi functionality by detecting the rssi source set in the FC. Only sends rssi if source is NONE or MSP. Rssi also broken out into it's own file. Unloaded when not needed.

Changed rssi scaling. It is now 0 - 99 to 0 - 255. Chose to do it like this because it doesn't need changes in the firmware and it's backwards compatible with earlier versions. No need for the user to adjust scaling/offset etc.

The INTERVAL is now separate for data_init and rssi. This gives better control over what parts of the code to run at that interval. Previously the buffer was polled at the same interval as the commands were sent. Now it's polled every lua cycle. This speeds up apiVersion detection by up to 0.5 seconds. Same for RTC sync.

Renamed rssi.lua function script to bfbkgd.lua(sorry. 6 char limit). Used "bf" to show where the script comes from to avoid confusion as people might have scripts from different sources in the FUNCTION folder.