VTX Table download
Downloads vtx tables if  supported and available. This is done as part of ui.lua's init state.

The vtx tables are saved as /BF/VTX/modelname.lua where "modelname" is the name of the currently selected model. Delete this file if the vtx tables need to be updated for some reason. The script will download them again.