Clear LCD for CMS script at startup
Clears the LCD and draws the "refresh[whateverbuttonrefreshis]" at the top row when the script is launched.

The current behaviour is to not clear the screen when the script is launched. If it's not responding and a manual refresh is required, there's no way for the user to know this. By clearing the screen and showing the message, the user should get an idea of what has to be done.